+++ app/vlstorage/netinsert/netinsert.go
@@ -18,10 +18,10 @@ import (
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/logger"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/promauth"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/timerpool"
+	"github.com/VictoriaMetrics/metrics"
 	"github.com/valyala/fastrand"
 
 	"github.com/VictoriaMetrics/VictoriaLogs/lib/logstorage"
-	"github.com/VictoriaMetrics/metrics"
 )
 
 // the maximum size of a single data block sent to storage node.
@@ -70,9 +70,11 @@ type storageNode struct {
 	// sendErrors counts failed send attempts for this storage node.
 	sendErrors *metrics.Counter
 
-	// the unix timestamp until the storageNode is disabled for data writing.
+	// disabledUntil contains unix timestamp until the storageNode is disabled for data writing.
 	disabledUntil atomic.Uint64
-	isBroken      atomic.Uint32
+
+	// isReachable is set to true if the given storageNode is available for data writing.
+	isReachable atomic.Bool
 }
 
 func newStorageNode(s *Storage, addr string, ac *promauth.Config, isTLS bool) *storageNode {
@@ -94,10 +96,12 @@ func newStorageNode(s *Storage, addr string, ac *promauth.Config, isTLS bool) *s
 		},
 		ac: ac,
 
-		sendErrors: metrics.GetOrCreateCounter(fmt.Sprintf(`vl_insert_remote_send_errors_total{url=%q}`, addr)),
+		sendErrors: metrics.GetOrCreateCounter(fmt.Sprintf(`vl_insert_remote_send_errors_total{addr=%q}`, addr)),
 
 		pendingData: &bytesutil.ByteBuffer{},
 	}
+
+	sn.isReachable.Store(true)
 
 	s.wg.Add(1)
 	go func() {
@@ -105,11 +109,11 @@ func newStorageNode(s *Storage, addr string, ac *promauth.Config, isTLS bool) *s
 		sn.backgroundFlusher()
 	}()
 
-	_ = metrics.GetOrCreateGauge(fmt.Sprintf(`vl_insert_remote_is_reachable{url=%q}`, addr), func() float64 {
-		if sn.isBroken.Load() == 1 {
-			return 0
+	_ = metrics.GetOrCreateGauge(fmt.Sprintf(`vl_insert_remote_is_reachable{addr=%q}`, addr), func() float64 {
+		if sn.isReachable.Load() {
+			return 1
 		}
-		return 1
+		return 0
 	})
 
 	return sn
@@ -196,7 +200,7 @@ func (sn *storageNode) mustSendInsertRequest(pendingData *bytesutil.ByteBuffer) 
 		logger.Warnf("%s; re-routing the data block to the remaining nodes", err)
 	}
 	for !sn.s.sendInsertRequestToAnyNode(pendingData) {
-		logger.Errorf("cannot send pending data to all storage nodes, since all of them are unavailable; re-trying to send the data in a second")
+		logger.Errorf("cannot send pending data to storage nodes, since all of them are unavailable; re-trying to send the data in a second")
 
 		t := timerpool.Get(time.Second)
 		select {
@@ -246,6 +250,7 @@ func (sn *storageNode) sendInsertRequest(pendingData *bytesutil.ByteBuffer) erro
 		req.Header.Set("Content-Encoding", "zstd")
 	}
 	if err := sn.ac.SetHeaders(req, true); err != nil {
+		sn.sendErrors.Inc()
 		return fmt.Errorf("cannot set auth headers for %q: %w", reqURL, err)
 	}
 
@@ -257,7 +262,7 @@ func (sn *storageNode) sendInsertRequest(pendingData *bytesutil.ByteBuffer) erro
 	defer resp.Body.Close()
 
 	if resp.StatusCode/100 == 2 {
-		sn.isBroken.Store(0)
+		sn.isReachable.Store(true)
 		return nil
 	}
 
@@ -276,9 +281,11 @@ func (sn *storageNode) getRequestURL(path string) string {
 }
 
 func (sn *storageNode) setDisableTemporarily() {
+	// Disable sending data to this sn for 10 seconds.
 	sn.disabledUntil.Store(fasttime.UnixTimestamp() + 10)
-	sn.isBroken.Store(1)
+
 	sn.sendErrors.Inc()
+	sn.isReachable.Store(false)
 }
 
 var zstdBufPool bytesutil.ByteBufferPool